backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1

# Supabase
supabase==2.11.0

# HTTP & Crawling
httpx==0.28.0
beautifulsoup4==4.12.3
lxml==5.3.0

# AI / LLM
google-genai==1.12.0

# Auth
PyJWT[crypto]==2.9.0

# Logging
structlog==24.4.0

# Utilities
python-multipart==0.0.18
